requirements.txt
attrs==25.1.0
aws-cdk-lib==2.180.0
aws-cdk.asset-awscli-v1==2.2.224
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==39.2.20
boto3==1.36.26
botocore==1.36.26
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
constructs==10.4.2
exceptiongroup==1.2.2
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.107.0
numpy==2.0.2
pandas==2.2.3
publication==0.0.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
SQLAlchemy==2.0.38
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
zipp==3.21.0
